+++ src/Telegram.Bot/Exceptions/ExceptionParser.cs
@@ -1,0 +1,13 @@
+using Telegram.Bot.Types;
+
+namespace Telegram.Bot.Exceptions
+{
+    internal class ExceptionParser : IExceptionParser
+    {
+        public ApiRequestException Parse(
+            int errorCode,
+            string description,
+            ResponseParameters? responseParameters) =>
+            new ApiRequestException(description, errorCode, responseParameters);
+    }
+}