+++ src/Telegram.Bot/Helpers/Extensions.cs
@@ -1,8 +1,13 @@
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Text;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
+using Telegram.Bot.Exceptions;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 
@@ -55,6 +60,92 @@ namespace Telegram.Bot.Helpers
                     multipartContent.AddStreamContent(mediaThumb.Content!, mediaThumb.FileName!);
                 }
             }
+        }
+
+        /// <summary>
+        /// Deserialized JSON in Stream into <typeparamref name="T"/>
+        /// </summary>
+        /// <param name="stream"><see cref="Stream"/> with content</param>
+        /// <typeparam name="T">Type of the resulting object</typeparam>
+        /// <returns></returns>
+        [return: MaybeNull]
+        private static T? DeserializeJsonFromStream<T>(this Stream stream)
+            where T : class
+        {
+            if (stream == null || stream.CanRead == false)
+                return default;
+
+            using var streamReader = new StreamReader(stream);
+            using var jsonTextReader = new JsonTextReader(streamReader);
+
+            var jsonSerializer = new JsonSerializer();
+            var searchResult = jsonSerializer.Deserialize<T>(jsonTextReader);
+
+            return searchResult;
+        }
+
+        /// <summary>
+        /// Deserialize body from HttpContent into <typeparamref name="T"/>
+        /// </summary>
+        /// <param name="httpResponse"><see cref="HttpResponseMessage"/> instance</param>
+        /// <param name="statusCode"><see cref="HttpStatusCode"/> of the response</param>
+        /// <typeparam name="T">Type of the resulting object</typeparam>
+        /// <returns></returns>
+        /// <exception cref="RequestException">
+        /// Thrown when body in the response can not be deserialized into <typeparamref name="T"/>
+        /// </exception>
+        internal static async Task<T> DeserializeContentAsync<T>(
+            this HttpResponseMessage httpResponse,
+            HttpStatusCode statusCode)
+            where T : class
+        {
+            T? deserializedObject;
+            Stream? contentStream = null;
+
+            if (httpResponse.Content is null)
+                throw new RequestException("Response doesn't contain any content", statusCode);
+
+            try
+            {
+                contentStream = await httpResponse.Content
+                    .ReadAsStreamAsync()
+                    .ConfigureAwait(false);
+
+                deserializedObject = contentStream
+                    .DeserializeJsonFromStream<T>();
+            }
+            catch (Exception exception)
+            {
+                var stringifiedResponse = await httpResponse.Content
+                    .ReadAsStringAsync()
+                    .ConfigureAwait(false);
+
+                throw new RequestException(
+                    "Required properties not found in response.",
+                    statusCode,
+                    stringifiedResponse,
+                    exception
+                );
+            }
+            finally
+            {
+                contentStream?.Dispose();
+            }
+
+            if (deserializedObject is null)
+            {
+                var stringifiedResponse = await httpResponse.Content
+                    .ReadAsStringAsync()
+                    .ConfigureAwait(false);
+
+                throw new RequestException(
+                    "Required properties not found in response.",
+                    statusCode,
+                    stringifiedResponse
+                );
+            }
+
+            return deserializedObject;
         }
     }
 }