+++ test/Telegram.Bot.Tests.Integ/Games/GamesFixture.cs
@@ -3,13 +3,14 @@ using System.Threading.Tasks;
 using Telegram.Bot.Exceptions;
 using Telegram.Bot.Tests.Integ.Framework;
 using Telegram.Bot.Types;
+using Xunit;
 
 namespace Telegram.Bot.Tests.Integ.Games
 {
     /// <summary>
     /// This fixture is for Games tests and it ensures test bot has a game set up before running the test methods.
     /// </summary>
-    public class GamesFixture
+    public class GamesFixture : IAsyncLifetime
     {
         public string GameShortName { get; }
 
@@ -30,24 +31,6 @@ namespace Telegram.Bot.Tests.Integ.Games
         {
             _fixture = fixture;
             GameShortName = "game1";
-            InitAsync().GetAwaiter().GetResult();
-        }
-
-        private async Task InitAsync()
-        {
-            try
-            {
-                await _fixture.BotClient.SendGameAsync(_fixture.SupergroupChat.Id, GameShortName);
-            }
-            catch (InvalidGameShortNameException e)
-            {
-                throw new ArgumentException(
-                    $@"Bot doesn't have game: ""{GameShortName}"". Make sure you set up a game with @BotFather.",
-                    e.Parameter, e
-                );
-            }
-
-            Player = await GetPlayerIdFromChatAdmins(_fixture.SupergroupChat.Id);
         }
 
         private async Task<User> GetPlayerIdFromChatAdmins(long chatId)
@@ -56,5 +39,24 @@ namespace Telegram.Bot.Tests.Integ.Games
             ChatMember player = admins[new Random(DateTime.Now.Millisecond).Next(admins.Length)];
             return player.User;
         }
+
+        public async Task InitializeAsync()
+        {
+            try
+            {
+                await _fixture.BotClient.SendGameAsync(_fixture.SupergroupChat.Id, GameShortName);
+            }
+            catch (ApiRequestException e) when (e.ErrorCode == 400 && e.Message.Contains("game_short_name"))
+            {
+                throw new ArgumentException(
+                    $@"Bot doesn't have game: ""{GameShortName}"". Make sure you set up a game with @BotFather.",
+                    e
+                );
+            }
+
+            Player = await GetPlayerIdFromChatAdmins(_fixture.SupergroupChat.Id);
+        }
+
+        public Task DisposeAsync() => Task.CompletedTask;
     }
 }