+++ CHANGELOG.md
@@ -26,6 +26,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Constructor to `ApiRequestEventArgs`
 - Constructor to `ApiRequestEventArgs`
 - Property `bool IRequest.IsWebhookResponse`
+- New base exception type `RequestException`
+- Method `ITelegramBotClient.SendRequestAsync` which returns `ApiResponse<T>` as is without throwing
+exception `ApiRequestException`
+- Interface `IExceptionParser` for implementing custom parsers for exception, default
+implementation of `IExceptionParser` throws all correct errors from Bot Api as `ApiRequestException`
+- Property `ITelegramBotClient.ExceptionParser` 
 
 ### Changed
 - All optional Bot API properties are made nullable 
@@ -49,6 +55,12 @@ feature)
 - Parameter order in `SendVideoNoteAsync`
 - Parameter order in `SendVenueAsync`
 - Parameter order in `SendContactAsync`
+- `ApiRequestException` extends `RequestException`
+- All error responses from Bot API which contain correct JSON string are now thrown as
+`ApiRequestException` from `ITelegramBotClient.MakeRequestException`
+- All other exceptions during request sending logic are wrapped inside `RequestException` 
+except `TaskCancelledException` that is triggered by token cancellation and errors from 
+`MakingApiRequest` and `ApiResponseReceived` event handlers
 
 ### Removed
 - Enum member `ParseMode.Default`
@@ -84,7 +96,8 @@ feature)
     - Event `OnReceiveError`
     - Event `OnReceiveGeneralError`
     - Method `StartReceiving`
-    - Method `StopReceiving` 
+    - Method `StopReceiving`
+- All exception types except `ApiRequestException`
 
 ## [15.7.1] - 2020-06-18